requirements.txt
asgiref==3.8.1
Django==5.0.6
django-debug-toolbar==4.4.3
djangorestframework==3.15.2
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.12.2
